chore: dialog mixin moved to widgets directory
additional:
- directory name should be lowercase
